Merge Dev to Master

+++ setupAndRunBTABackend.sh
@@ -7,11 +7,54 @@ MAC_OS="darwin-amd64"
 LINUX_OS="linux-amd64"
 ARCH=$(echo "$(uname -s|tr '[:upper:]' '[:lower:]'|sed 's/mingw64_nt.*/windows/')-$(uname -m |sed 's/x86_64/amd64/g')" |sed 's/darwin-arm64/darwin-amd64/g')
 
+which ifconfig
+if [ "$?" -ne 0 ]; then
+    if [ "$ARCH" = "$LINUX_OS" ]; then
+        sudo apt-get update && sudo apt-get install net-tools
+    else
+        echo
+        echo -e "${RED}"
+        echo "-----------------------------------------------------------------------------------------"
+        echo "-----------------------------------------------------------------------------------------"
+        echo "OS Not Found. Please install ifconfig manually on your device and re-run the script"
+        echo "-----------------------------------------------------------------------------------------"
+        echo "-----------------------------------------------------------------------------------------"
+        echo -e "${COLOR_OFF}"
+        exit 1
+    fi
+fi
+
+which ifconfig
+if [ "$?" -ne 0 ]; then
+    echo
+    echo -e "${RED}"
+    echo "-----------------------------------------------------------------------------------------"
+    echo "-----------------------------------------------------------------------------------------"
+    echo "Failed to install ifconfig. Please install ifconfig manually on your device and re-run the script"
+    echo "-----------------------------------------------------------------------------------------"
+    echo "-----------------------------------------------------------------------------------------"
+    echo -e "${COLOR_OFF}"
+    exit 1
+fi
+
 # Getting IP Address For Blockchain Network
 export PRIVATE_NETWORK_IP_ADDRESS=$(ifconfig | grep -E "([0-9]{1,3}\.){3}[0-9]{1,3}" | grep -v 127.0.0.1 | tail -1 | awk '{ print $2 }')
 
+if [[ ! -f ".env" ||  ! -s ".env" ]];
+then
+echo -e "${GREEN}"
+echo -e "------------------------------------------------------------------------------------------------------"
+echo -e "Creating .env file"
+
 # Copy the env sample and create new .env file and paste all the contents there.
 cp -r env-samples/.env.local.sample .env
+
+echo -e ""
+echo -e "Successfully created .env file"
+echo "------------------------------------------------------------------------------------------------------"
+echo -e "${Color_Off}"
+fi
+
 
 if [ "$ARCH" = "$MAC_OS" ];
 then
@@ -26,40 +69,11 @@ fi
 
 export APP_RUNNED_SUCCESS_RESPONSE="APP IS LISTENING TO PORT ${PORT}"
 
-echo "Pulling latest code"
+. ./scripts/create-default-directories.sh
 
+echo "Pulling latest code"
 git pull
 
-if [ -d "$DATABASE_VOLUME_MOUNT" ]; then
-    echo "Database Directory already exists"
-    echo "Changing ownership to current user"
-    sudo chown -R $USER:$USER $DATABASE_VOLUME_MOUNT
-else
-    echo "Creating Database Directory and Changing ownership to current user"
-    mkdir -p $DATABASE_VOLUME_MOUNT && chown -R $USER:$USER $DATABASE_VOLUME_MOUNT
-fi
-
-LOGS_DIR=./logs
-
-if [ -d "$LOGS_DIR" ]; then
-    echo "Logs Directory already exists"
-    echo "Changing ownership to current user"
-    sudo chown -R $USER:$USER $LOGS_DIR
-else
-    echo "Creating Logs Directory and Changing ownership to current user"
-    mkdir -p $LOGS_DIR && chown -R $USER:$USER $LOGS_DIR
-fi
-
-UPLOADS_DIR=./uploads
-
-if [ -d $UPLOADS_DIR ]; then
-    echo "Uploads Directory already exists"
-    echo "Changing ownership to current user"
-    sudo chown -R $USER:$USER $UPLOADS_DIR
-else
-    echo "Creating Uploads Directory and Changing ownership to current user"
-    mkdir -p $UPLOADS_DIR && chown -R $USER:$USER $UPLOADS_DIR
-fi
 
 . ./scripts/docker-down.sh
 
@@ -91,7 +105,7 @@ fi
 
 echo -e "${GREEN}"
 echo "--------------------------------------------------------------------"
-echo "Please wait while bta-backend application is completely started..."
+echo "Please wait while bta-backend application is completely started retrying($[$COUNTER +1])..."
 echo "--------------------------------------------------------------------"
 echo -e "${COLOR_OFF}"
 
@@ -102,36 +116,11 @@ GET_APP_RUNNING_STATUS=$(docker logs bta_api_local 2>&1 | grep "$APP_RUNNED_SUCC
 
 done
 
-echo -e "${GREEN}"
-echo "----------------------------------------------------------"
-echo "Starting Migrating App Migration"
-echo "----------------------------------------------------------"
-echo -e "${COLOR_OFF}"
-
 # Run App Migrations
 . ./scripts/app-migration.sh
 
-echo -e "${GREEN}"
-echo "----------------------------------------------------------"
-echo "Successfully Migrated App Migration"
-echo "----------------------------------------------------------"
-echo -e "${COLOR_OFF}"
-
-echo -e "${GREEN}"
-echo "----------------------------------------------------------"
-echo "Starting Migrating Super Admin Migration"
-echo "----------------------------------------------------------"
-echo -e "${COLOR_OFF}"
-
 # Run Super Admin Migrations
 . ./scripts/super-admin-migration.sh
-
-echo -e "${GREEN}"
-echo "----------------------------------------------------------"
-
-echo "Successfully Migrated Super Admin Migration"
-echo "----------------------------------------------------------"
-echo -e "${COLOR_OFF}"
 
 echo -e "${GREEN}"
 echo "----------------------------------------------------------"
